move common methods to utility obdUtils.py
