+++ imuSensor.py
@@ -6,6 +6,8 @@ import board
 import adafruit_bno055
 import socketio
 
+import obdUtils
+
 SENSOR_TYPE = 'IMU'
 ERROR = 'ERR'
 INFO = 'INFO'
@@ -19,21 +21,7 @@ sensor = adafruit_bno055.BNO055_I2C(i2c)
 last_val = 0xFFFF
 
 sio = None
-
-
 numTries = 1
-
-def createLogMessage(logType, sensor, description, detailedDescription):
-    now = datetime.datetime.now()
-    timestamp = "%d:%d:%d" % (now.hour, now.minute, now.second)
-            
-    return {
-        'logType': logType,
-        'sensor' : SENSOR_TYPE,
-        'exType' : description,
-        'args' : detailedDescription,
-        'timeStamp': timestamp
-    }
 
 
 def temperature():
@@ -87,7 +75,7 @@ def emitImuData():
             
         except Exception as ex: #logs any errors encountered during reading of gps. Also allows program to pick back up if node server connection is lost
                 
-                errorLog = createLogMessage(ERROR, SENSOR_TYPE, type(ex).__name__, ex.args)
+                errorLog = obdUtils.createLogMessage(ERROR, SENSOR_TYPE, type(ex).__name__, ex.args)
                 print(errorLog)
                 sio.emit('log', json.dumps(errorLog)) #will only work if exception is unrelated to node server connection
                 i2c = board.I2C() #try to reconnect to BNO055 sensor
@@ -107,7 +95,7 @@ while True: #loop until a connection is made with the server instead of immediat
     except Exception as ex:
         numTries += 1
         print("IMU app unable to connect to node server, retrying attempt {0}".format(numTries))
-        errorLog = createLogMessage(ERROR, SENSOR_TYPE, type(ex).__name__, ex.args)
+        errorLog = obdUtils.createLogMessage(ERROR, SENSOR_TYPE, type(ex).__name__, ex.args)
         print(errorLog)
         time.sleep(1)
             
@@ -116,7 +104,7 @@ while True: #loop until a connection is made with the server instead of immediat
 @sio.event
 def connect():
     
-    connectedLog = createLogMessage(INFO, SENSOR_TYPE, 'Connection', 'Established')
+    connectedLog = obdUtils.createLogMessage(INFO, SENSOR_TYPE, 'Connection', 'Established')
     print(connectedLog)
     sio.emit('log', json.dumps(connectedLog))
